+++ client/src/components/bottomNavigator/BottomNavigator.style.ts
